fix: main.goの配置先移動
